added new daemon
