+++ sonic-xcvrd/tests/test_xcvrd.py
@@ -21,9 +21,6 @@ swsscommon.SubscriberStateTable = MagicMock()
 swsscommon.SonicDBConfig = MagicMock()
 #swsscommon.Select = MagicMock()
 
-sys.modules['sonic_y_cable'] = MagicMock()
-sys.modules['sonic_y_cable.y_cable'] = MagicMock()
-
 test_path = os.path.dirname(os.path.abspath(__file__))
 modules_path = os.path.dirname(test_path)
 scripts_path = os.path.join(modules_path, "xcvrd")
@@ -31,7 +28,6 @@ sys.path.insert(0, modules_path)
 
 os.environ["XCVRD_UNIT_TESTING"] = "1"
 from xcvrd.xcvrd import *
-from xcvrd.xcvrd_utilities.y_cable_helper import *
 from xcvrd.xcvrd_utilities.sfp_status_helper import *
 from xcvrd.xcvrd_utilities.port_mapping import *
 
@@ -43,8 +39,6 @@ global_media_settings = media_settings_with_comma_dict['GLOBAL_MEDIA_SETTINGS'].
 media_settings_with_comma_dict['GLOBAL_MEDIA_SETTINGS']['1-5,6,7-20,21-32'] = global_media_settings
 
 class TestXcvrdScript(object):
-    def test_xcvrd_helper_class_run(self):
-        Y_cable_task = YCableTableUpdateTask(None)
 
     @patch('xcvrd.xcvrd._wrapper_get_sfp_type')
     @patch('xcvrd.xcvrd_utilities.port_mapping.PortMapping.logical_port_name_to_physical_port_list', MagicMock(return_value=[0]))
@@ -239,148 +233,6 @@ class TestXcvrdScript(object):
         stop_event = threading.Event()
         init_port_sfp_status_tbl(port_mapping, stop_event)
 
-    @patch('xcvrd.xcvrd_utilities.y_cable_helper.y_cable_platform_sfputil', MagicMock(return_value=[0]))
-    @patch('xcvrd.xcvrd_utilities.y_cable_helper.y_cable_wrapper_get_presence', MagicMock(return_value=True))
-    @patch('xcvrd.xcvrd_utilities.port_mapping.PortMapping.logical_port_name_to_physical_port_list', MagicMock(return_value=[0]))
-    @patch('xcvrd.xcvrd_utilities.y_cable_helper.get_muxcable_info', MagicMock(return_value={'tor_active': 'self',
-                                                                       'mux_direction': 'self',
-                                                                       'manual_switch_count': '7',
-                                                                       'auto_switch_count': '71',
-                                                                       'link_status_self': 'up',
-                                                                       'link_status_peer': 'up',
-                                                                       'link_status_nic': 'up',
-                                                                       'nic_lane1_active': 'True',
-                                                                       'nic_lane2_active': 'True',
-                                                                       'nic_lane3_active': 'True',
-                                                                       'nic_lane4_active': 'True',
-                                                                       'self_eye_height_lane1': '500',
-                                                                       'self_eye_height_lane2': '510',
-                                                                       'peer_eye_height_lane1': '520',
-                                                                       'peer_eye_height_lane2': '530',
-                                                                       'nic_eye_height_lane1': '742',
-                                                                       'nic_eye_height_lane2': '750',
-                                                                       'internal_temperature': '28',
-                                                                       'internal_voltage': '3.3',
-                                                                       'nic_temperature': '20',
-                                                                       'nic_voltage': '2.7',
-                                                                       'version_nic_active': '1.6MS',
-                                                                       'version_nic_inactive': '1.7MS',
-                                                                       'version_nic_next': '1.7MS',
-                                                                       'version_self_active': '1.6MS',
-                                                                       'version_self_inactive': '1.7MS',
-                                                                       'version_self_next': '1.7MS',
-                                                                       'version_peer_active': '1.6MS',
-                                                                       'version_peer_inactive': '1.7MS',
-                                                                       'version_peer_next': '1.7MS'}))
-    def test_post_port_mux_info_to_db(self):
-        logical_port_name = "Ethernet0"
-        port_mapping = PortMapping()
-        mux_tbl = Table("STATE_DB", y_cable_helper.MUX_CABLE_INFO_TABLE)
-        rc = post_port_mux_info_to_db(logical_port_name, port_mapping, mux_tbl)
-        assert(rc != -1)
-
-
-    @patch('xcvrd.xcvrd_utilities.y_cable_helper.y_cable_platform_sfputil', MagicMock(return_value=[0]))
-    @patch('xcvrd.xcvrd_utilities.y_cable_helper.y_cable_wrapper_get_presence', MagicMock(return_value=True))
-    @patch('xcvrd.xcvrd_utilities.port_mapping.PortMapping.logical_port_name_to_physical_port_list', MagicMock(return_value=[0]))
-    @patch('xcvrd.xcvrd_utilities.y_cable_helper.get_muxcable_static_info', MagicMock(return_value={'read_side': 'self',
-                                                                              'nic_lane1_precursor1': '1',
-                                                                              'nic_lane1_precursor2': '-7',
-                                                                              'nic_lane1_maincursor': '-1',
-                                                                              'nic_lane1_postcursor1': '11',
-                                                                              'nic_lane1_postcursor2': '11',
-                                                                              'nic_lane2_precursor1': '12',
-                                                                              'nic_lane2_precursor2': '7',
-                                                                              'nic_lane2_maincursor': '7',
-                                                                              'nic_lane2_postcursor1': '7',
-                                                                              'nic_lane2_postcursor2': '7',
-                                                                              'tor_self_lane1_precursor1': '17',
-                                                                              'tor_self_lane1_precursor2': '17',
-                                                                              'tor_self_lane1_maincursor': '17',
-                                                                              'tor_self_lane1_postcursor1': '17',
-                                                                              'tor_self_lane1_postcursor2': '17',
-                                                                              'tor_self_lane2_precursor1': '7',
-                                                                              'tor_self_lane2_precursor2': '7',
-                                                                              'tor_self_lane2_maincursor': '7',
-                                                                              'tor_self_lane2_postcursor1': '7',
-                                                                              'tor_self_lane2_postcursor2': '7',
-                                                                              'tor_peer_lane1_precursor1': '7',
-                                                                              'tor_peer_lane1_precursor2': '7',
-                                                                              'tor_peer_lane1_maincursor': '17',
-                                                                              'tor_peer_lane1_postcursor1': '7',
-                                                                              'tor_peer_lane1_postcursor2': '17',
-                                                                              'tor_peer_lane2_precursor1': '7',
-                                                                              'tor_peer_lane2_precursor2': '7',
-                                                                              'tor_peer_lane2_maincursor': '17',
-                                                                              'tor_peer_lane2_postcursor1': '7',
-                                                                              'tor_peer_lane2_postcursor2': '17'}))
-    def test_post_port_mux_static_info_to_db(self):
-        logical_port_name = "Ethernet0"
-        port_mapping = PortMapping()
-        mux_tbl = Table("STATE_DB", y_cable_helper.MUX_CABLE_STATIC_INFO_TABLE)
-        rc = post_port_mux_static_info_to_db(logical_port_name, port_mapping, mux_tbl)
-        assert(rc != -1)
-
-    def test_y_cable_helper_format_mapping_identifier1(self):
-        rc = format_mapping_identifier("ABC        ")
-        assert(rc == "abc")
-
-    def test_y_cable_wrapper_get_transceiver_info(self):
-        with patch('xcvrd.xcvrd_utilities.y_cable_helper.y_cable_platform_sfputil') as patched_util:
-            patched_util.get_transceiver_info_dict.return_value = {'manufacturer': 'Microsoft',
-                                                                              'model': 'model1'}
-
-            transceiver_dict = y_cable_wrapper_get_transceiver_info(1)
-            vendor = transceiver_dict.get('manufacturer')
-            model = transceiver_dict.get('model')
-
-        assert(vendor == "Microsoft")
-        assert(model == "model1")
-
-    def test_y_cable_wrapper_get_presence(self):
-        with patch('xcvrd.xcvrd_utilities.y_cable_helper.y_cable_platform_sfputil') as patched_util:
-            patched_util.get_presence.return_value = True
-
-            presence = y_cable_wrapper_get_presence(1)
-
-        assert(presence == True)
-
-    @patch('xcvrd.xcvrd_utilities.port_mapping.PortMapping.logical_port_name_to_physical_port_list', MagicMock(return_value=[0]))
-    @patch('xcvrd.xcvrd_utilities.y_cable_helper.y_cable_wrapper_get_presence', MagicMock(return_value=True))
-    def test_get_ycable_physical_port_from_logical_port(self):
-        port_mapping = PortMapping()
-        instance = get_ycable_physical_port_from_logical_port("Ethernet0", port_mapping)
-
-        assert(instance == 0)
-
-    @patch('xcvrd.xcvrd_utilities.port_mapping.PortMapping.logical_port_name_to_physical_port_list', MagicMock(return_value=[0]))
-    @patch('xcvrd.xcvrd_utilities.y_cable_helper.y_cable_wrapper_get_presence', MagicMock(return_value=True))
-    def test_get_ycable_port_instance_from_logical_port(self):
-
-        with patch('xcvrd.xcvrd_utilities.y_cable_helper.y_cable_port_instances') as patched_util:
-            patched_util.get.return_value = 0
-            port_mapping = PortMapping()
-            instance = get_ycable_port_instance_from_logical_port("Ethernet0", port_mapping)
-
-        assert(instance == 0)
-
-    def test_set_show_firmware_fields(self):
-
-        mux_info_dict = {}
-        xcvrd_show_fw_res_tbl = Table("STATE_DB", "XCVRD_SHOW_FW_RES")
-        mux_info_dict['version_self_active'] = '0.8'
-        mux_info_dict['version_self_inactive'] = '0.7'
-        mux_info_dict['version_self_next'] = '0.7'
-        mux_info_dict['version_peer_active'] = '0.8'
-        mux_info_dict['version_peer_inactive'] = '0.7'
-        mux_info_dict['version_peer_next'] = '0.7'
-        mux_info_dict['version_nic_active'] = '0.8'
-        mux_info_dict['version_nic_inactive'] = '0.7'
-        mux_info_dict['version_nic_next'] = '0.7'
-        rc = set_show_firmware_fields("Ethernet0", mux_info_dict, xcvrd_show_fw_res_tbl)
-
-        assert(rc == 0)
-
     def test_get_media_settings_key(self):
         xcvr_info_dict = {
             0: {
@@ -553,7 +405,7 @@ class TestXcvrdScript(object):
     def test_DomInfoUpdateTask_task_run_stop(self):
         port_mapping = PortMapping()
         task = DomInfoUpdateTask(port_mapping)
-        task.task_run([False])
+        task.task_run()
         task.task_stop()
         assert not task.task_thread.is_alive()
 
@@ -574,7 +426,7 @@ class TestXcvrdScript(object):
         task = DomInfoUpdateTask(port_mapping)
         task.task_stopping_event.wait = MagicMock(side_effect=[False, True])
         mock_detect_error.return_value = True
-        task.task_worker([False])
+        task.task_worker()
         assert task.port_mapping.logical_port_list.count('Ethernet0')
         assert task.port_mapping.get_asic_id_for_logical_port('Ethernet0') == 0
         assert task.port_mapping.get_physical_to_logical(1) == ['Ethernet0']
@@ -583,7 +435,7 @@ class TestXcvrdScript(object):
         assert mock_post_dom_info.call_count == 0
         mock_detect_error.return_value = False
         task.task_stopping_event.wait = MagicMock(side_effect=[False, True])
-        task.task_worker([False])
+        task.task_worker()
         assert mock_post_dom_th.call_count == 1
         assert mock_post_dom_info.call_count == 1
 
@@ -603,7 +455,7 @@ class TestXcvrdScript(object):
         port_change_event = PortChangeEvent('Ethernet0', 1, 0, PortChangeEvent.PORT_ADD)
         wait_time = 5
         while wait_time > 0:
-            task.on_port_config_change(stopping_event, [False], port_change_event)
+            task.on_port_config_change(stopping_event, port_change_event)
             if task.port_mapping.logical_port_list:
                 break
             wait_time -= 1
@@ -616,7 +468,7 @@ class TestXcvrdScript(object):
         port_change_event = PortChangeEvent('Ethernet0', 1, 0, PortChangeEvent.PORT_REMOVE)
         wait_time = 5
         while wait_time > 0:
-            task.on_port_config_change(stopping_event, [False], port_change_event)
+            task.on_port_config_change(stopping_event , port_change_event)
             if not task.port_mapping.logical_port_list:
                 break
             wait_time -= 1
@@ -631,7 +483,7 @@ class TestXcvrdScript(object):
         retry_eeprom_set = set()
         task = SfpStateUpdateTask(port_mapping, retry_eeprom_set)
         sfp_error_event = multiprocessing.Event()
-        task.task_run(sfp_error_event, [False])
+        task.task_run(sfp_error_event)
         assert wait_until(5, 1, task.task_process.is_alive)
         task.task_stop()
         assert wait_until(5, 1, lambda: task.task_process.is_alive() is False)
@@ -702,7 +554,7 @@ class TestXcvrdScript(object):
         mock_mapping_event.return_value = SYSTEM_NOT_READY
 
         # Test state machine: STATE_INIT + SYSTEM_NOT_READY event => STATE_INIT + SYSTEM_NOT_READY event ... => STATE_EXIT
-        task.task_worker(stop_event, sfp_error_event, [False])
+        task.task_worker(stop_event, sfp_error_event)
         assert mock_os_kill.call_count == 1
         assert sfp_error_event.is_set()
 
@@ -710,7 +562,7 @@ class TestXcvrdScript(object):
         mock_os_kill.reset_mock()
         sfp_error_event.clear()
         # Test state machine: STATE_INIT + SYSTEM_FAIL event => STATE_INIT + SYSTEM_FAIL event ... => STATE_EXIT
-        task.task_worker(stop_event, sfp_error_event, [False])
+        task.task_worker(stop_event, sfp_error_event)
         assert mock_os_kill.call_count == 1
         assert sfp_error_event.is_set()
 
@@ -718,7 +570,7 @@ class TestXcvrdScript(object):
         mock_os_kill.reset_mock()
         sfp_error_event.clear()
         # Test state machine: STATE_INIT + SYSTEM_BECOME_READY event => STATE_NORMAL + SYSTEM_NOT_READY event ... => STATE_EXIT
-        task.task_worker(stop_event, sfp_error_event, [False])
+        task.task_worker(stop_event, sfp_error_event)
         assert mock_os_kill.call_count == 1
         assert not sfp_error_event.is_set()
 
@@ -727,7 +579,7 @@ class TestXcvrdScript(object):
         sfp_error_event.clear()
         # Test state machine: STATE_INIT + SYSTEM_BECOME_READY event => STATE_NORMAL + SYSTEM_FAIL event ... => STATE_INIT
         # + SYSTEM_FAIL event ... => STATE_EXIT
-        task.task_worker(stop_event, sfp_error_event, [False])
+        task.task_worker(stop_event, sfp_error_event)
         assert mock_os_kill.call_count == 1
         assert sfp_error_event.is_set()
 
@@ -738,7 +590,7 @@ class TestXcvrdScript(object):
         mock_post_sfp_info.return_value = SFP_EEPROM_NOT_READY
         stop_event.is_set = MagicMock(side_effect=[False, True])
         # Test state machine: handle SFP insert event, but EEPROM read failure
-        task.task_worker(stop_event, sfp_error_event, [False])
+        task.task_worker(stop_event, sfp_error_event)
         assert mock_updata_status.call_count == 1
         assert mock_post_sfp_info.call_count == 2 # first call and retry call
         assert mock_post_dom_info.call_count == 0
@@ -752,7 +604,7 @@ class TestXcvrdScript(object):
         mock_updata_status.reset_mock()
         mock_post_sfp_info.reset_mock()
         # Test state machine: handle SFP insert event, and EEPROM read success
-        task.task_worker(stop_event, sfp_error_event, [False])
+        task.task_worker(stop_event, sfp_error_event)
         assert mock_updata_status.call_count == 1
         assert mock_post_sfp_info.call_count == 1
         assert mock_post_dom_info.call_count == 1
@@ -763,7 +615,7 @@ class TestXcvrdScript(object):
         mock_change_event.return_value = (True, {1:SFP_STATUS_REMOVED}, {})
         mock_updata_status.reset_mock()
         # Test state machine: handle SFP remove event
-        task.task_worker(stop_event, sfp_error_event, [False])
+        task.task_worker(stop_event, sfp_error_event)
         assert mock_updata_status.call_count == 1
         assert mock_del_dom.call_count == 1
 
@@ -773,7 +625,7 @@ class TestXcvrdScript(object):
         mock_updata_status.reset_mock()
         mock_del_dom.reset_mock()
         # Test state machine: handle SFP error event
-        task.task_worker(stop_event, sfp_error_event, [False])
+        task.task_worker(stop_event, sfp_error_event)
         assert mock_updata_status.call_count == 1
         assert mock_del_dom.call_count == 1
