+++ sonic-ycable/pytest.ini
@@ -1,0 +1,2 @@
+[pytest]
+addopts = --cov=ycable --cov-report html --cov-report term --cov-report xml --junitxml=test-results.xml -v
