+++ sonic-ycable/tests/mock_swsscommon.py
@@ -1,0 +1,31 @@
+STATE_DB = ''
+CHASSIS_STATE_DB = ''
+
+
+class Table:
+    def __init__(self, db, table_name):
+        self.table_name = table_name
+        self.mock_dict = {}
+
+    def _del(self, key):
+        if key in self.mock_dict:
+            del self.mock_dict[key]
+        pass
+
+    def set(self, key, fvs):
+        self.mock_dict[key] = fvs
+        pass
+
+    def get(self, key):
+        if key in self.mock_dict:
+            return self.mock_dict[key]
+        return None
+
+    def get_size(self):
+        return (len(self.mock_dict))
+
+
+class FieldValuePairs:
+    def __init__(self, fvs):
+        self.fv_dict = dict(fvs)
+        pass
