+++ sonic-ycable/tests/t0-sample-port-config.ini
@@ -1,0 +1,14 @@
+# name          lanes             alias
+Ethernet0       29,30,31,32       fortyGigE0/0
+Ethernet4       25,26,27,28       fortyGigE0/4
+Ethernet8       37,38,39,40       fortyGigE0/8
+Ethernet12      33,34,35,36       fortyGigE0/12
+Ethernet16      41,42,43,44       fortyGigE0/16
+Ethernet20      45,46,47,48       fortyGigE0/20
+Ethernet24      5,6,7,8           fortyGigE0/24
+Ethernet28      1,2,3,4           fortyGigE0/28
+Ethernet32      9,10,11,12        fortyGigE0/32
+Ethernet36      13,14,15,16       fortyGigE0/36
+Ethernet40      21,22,23,24       fortyGigE0/40
+Ethernet44      17,18,19,20       fortyGigE0/44
+Ethernet48      49,50,51,52       fortyGigE0/48
