+++ sonic-ycable/ycable/ycable_utilities/y_cable_helper.py
@@ -3,11 +3,11 @@
     helper utlities configuring y_cable for xcvrd daemon
 """
 
-import copy
 import datetime
 import os
 import re
 import threading
+import time
 
 from importlib import import_module
 
@@ -16,7 +16,6 @@ from sonic_py_common import multi_asic
 from sonic_y_cable import y_cable_vendor_mapping
 from swsscommon import swsscommon
 from . import sfp_status_helper
-from .port_mapping import read_port_config_change
 
 
 SELECT_TIMEOUT = 1000
@@ -79,6 +78,19 @@ def format_mapping_identifier(string):
 
     return s
 
+# Find out the underneath physical port list by logical name
+
+
+def logical_port_name_to_physical_port_list(port_name):
+    if port_name.startswith("Ethernet"):
+        if y_cable_platform_sfputil.is_logical_port(port_name):
+            return y_cable_platform_sfputil.get_logical_to_physical(port_name)
+        else:
+            helper_logger.log_error("Invalid port '%s'" % port_name)
+            return None
+    else:
+        return [int(port_name)]
+
 
 def y_cable_wrapper_get_presence(physical_port):
     if y_cable_platform_chassis is not None:
@@ -87,6 +99,7 @@ def y_cable_wrapper_get_presence(physical_port):
         except NotImplementedError:
             pass
     return y_cable_platform_sfputil.get_presence(physical_port)
+
 
 
 def hook_y_cable_simulated(target):
@@ -129,9 +142,9 @@ def y_cable_wrapper_get_transceiver_info(physical_port):
             pass
     return y_cable_platform_sfputil.get_transceiver_info_dict(physical_port)
 
-def get_ycable_physical_port_from_logical_port(logical_port_name, port_mapping):
+def get_ycable_physical_port_from_logical_port(logical_port_name):
 
-    physical_port_list = port_mapping.logical_port_name_to_physical_port_list(logical_port_name)
+    physical_port_list = logical_port_name_to_physical_port_list(logical_port_name)
 
     if len(physical_port_list) == 1:
 
@@ -152,9 +165,9 @@ def get_ycable_physical_port_from_logical_port(logical_port_name, port_mapping):
             "Error: Retreived multiple ports for a Y cable table port {} while retreiving physical port mapping".format(logical_port_name))
         return -1
 
-def get_ycable_port_instance_from_logical_port(logical_port_name, port_mapping):
+def get_ycable_port_instance_from_logical_port(logical_port_name):
 
-    physical_port_list = port_mapping.logical_port_name_to_physical_port_list(logical_port_name)
+    physical_port_list = logical_port_name_to_physical_port_list(logical_port_name)
 
     if len(physical_port_list) == 1:
 
@@ -265,8 +278,8 @@ def y_cable_toggle_mux_torB(physical_port):
         return -1
 
 
-def update_tor_active_side(read_side, state, logical_port_name, port_mapping):
-    physical_port_list = port_mapping.logical_port_name_to_physical_port_list(
+def update_tor_active_side(read_side, state, logical_port_name):
+    physical_port_list = logical_port_name_to_physical_port_list(
         logical_port_name)
 
     if len(physical_port_list) == 1:
@@ -300,14 +313,14 @@ def update_tor_active_side(read_side, state, logical_port_name, port_mapping):
         return -1
 
 
-def update_appdb_port_mux_cable_response_table(logical_port_name, port_mapping, asic_index, appl_db, read_side):
+def update_appdb_port_mux_cable_response_table(logical_port_name, asic_index, appl_db, read_side):
 
     status = None
     y_cable_response_tbl = {}
 
     y_cable_response_tbl[asic_index] = swsscommon.Table(
         appl_db[asic_index], "MUX_CABLE_RESPONSE_TABLE")
-    physical_port_list = port_mapping.logical_port_name_to_physical_port_list(
+    physical_port_list = logical_port_name_to_physical_port_list(
         logical_port_name)
 
     if len(physical_port_list) == 1:
@@ -376,8 +389,8 @@ def update_appdb_port_mux_cable_response_table(logical_port_name, port_mapping, 
             "Error: Retreived multiple ports for a Y cable port {} while responding to command probe".format(logical_port_name))
 
 
-def read_y_cable_and_update_statedb_port_tbl(logical_port_name, port_mapping, mux_config_tbl):
-    physical_port_list = port_mapping.logical_port_name_to_physical_port_list(
+def read_y_cable_and_update_statedb_port_tbl(logical_port_name, mux_config_tbl):
+    physical_port_list = logical_port_name_to_physical_port_list(
         logical_port_name)
 
     read_side = None
@@ -456,7 +469,7 @@ def read_y_cable_and_update_statedb_port_tbl(logical_port_name, port_mapping, mu
         helper_logger.log_warning(
             "Error: Retreived multiple ports for a Y cable port {} to perform read_y_cable update state db".format(logical_port_name))
 
-def create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, port_mapping):
+def create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name):
 
     namespaces = multi_asic.get_front_end_namespaces()
     for namespace in namespaces:
@@ -472,9 +485,9 @@ def create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_t
             state_db[asic_id], MUX_CABLE_INFO_TABLE)
     # fill the newly found entry
     read_y_cable_and_update_statedb_port_tbl(
-        logical_port_name, port_mapping, y_cable_tbl[asic_index])
+        logical_port_name, y_cable_tbl[asic_index])
 
-def check_identifier_presence_and_update_mux_table_entry(state_db, port_tbl, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, port_mapping, y_cable_presence):
+def check_identifier_presence_and_update_mux_table_entry(state_db, port_tbl, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, y_cable_presence):
 
     global y_cable_port_instances
     global y_cable_port_locks
@@ -491,10 +504,10 @@ def check_identifier_presence_and_update_mux_table_entry(state_db, port_tbl, y_c
 
             val = mux_table_dict.get("state", None)
 
-            if val in ["active", "auto"]:
+            if val in ["active", "auto", "manual", "standby"]:
 
                 # import the module and load the port instance
-                physical_port_list = port_mapping.logical_port_name_to_physical_port_list(
+                physical_port_list = logical_port_name_to_physical_port_list(
                     logical_port_name)
 
                 if len(physical_port_list) == 1:
@@ -509,7 +522,7 @@ def check_identifier_presence_and_update_mux_table_entry(state_db, port_tbl, y_c
                             if vendor is None:
                                 helper_logger.log_warning(
                                     "Error: Unable to find Vendor name for Transceiver for Y-Cable initiation {}".format(logical_port_name))
-                                create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, port_mapping)
+                                create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name)
                                 return
 
                             model = port_info_dict.get('model')
@@ -517,7 +530,7 @@ def check_identifier_presence_and_update_mux_table_entry(state_db, port_tbl, y_c
                             if model is None:
                                 helper_logger.log_warning(
                                     "Error: Unable to find model name for Transceiver for Y-Cable initiation {}".format(logical_port_name))
-                                create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, port_mapping)
+                                create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name)
                                 return
 
                             vendor = format_mapping_identifier(vendor)
@@ -527,14 +540,14 @@ def check_identifier_presence_and_update_mux_table_entry(state_db, port_tbl, y_c
                             if module_dir is None:
                                 helper_logger.log_warning(
                                     "Error: Unable to find module dir name from vendor for Y-Cable initiation {}".format(logical_port_name))
-                                create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, port_mapping)
+                                create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name)
                                 return
 
                             module = module_dir.get(model)
                             if module is None:
                                 helper_logger.log_warning(
                                     "Error: Unable to find module name from model for Y-Cable initiation {}".format(logical_port_name))
-                                create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, port_mapping)
+                                create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name)
                                 return
 
                             attr_name = 'sonic_y_cable.' + module
@@ -542,12 +555,12 @@ def check_identifier_presence_and_update_mux_table_entry(state_db, port_tbl, y_c
                                 y_cable_attribute = getattr(import_module(attr_name), 'YCable')
                             except Exception as e:
                                 helper_logger.log_warning("Failed to load the attr due to {}".format(repr(e)))
-                                create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, port_mapping)
+                                create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name)
                                 return
                             if y_cable_attribute is None:
                                 helper_logger.log_warning(
                                     "Error: Unable to import attr name for Y-Cable initiation {}".format(logical_port_name))
-                                create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, port_mapping)
+                                create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name)
                                 return
  
                             y_cable_port_instances[physical_port] = y_cable_attribute(physical_port, helper_logger)
@@ -557,13 +570,13 @@ def check_identifier_presence_and_update_mux_table_entry(state_db, port_tbl, y_c
                                     vendor_name_api = y_cable_port_instances.get(physical_port).get_vendor()
                                 except Exception as e:
                                     helper_logger.log_warning("Failed to call the get_vendor API for port {} due to {}".format(physical_port,repr(e)))
-                                    create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, port_mapping)
+                                    create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name)
                                     return
 
                             if format_mapping_identifier(vendor_name_api) != vendor:
                                 y_cable_port_instances.pop(physical_port)
                                 y_cable_port_locks.pop(physical_port)
-                                create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, port_mapping)
+                                create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name)
                                 helper_logger.log_warning("Error: Y Cable api does not work for {}, {} actual vendor name {}".format(
                                     logical_port_name, vendor_name_api, vendor))
                                 return
@@ -576,11 +589,11 @@ def check_identifier_presence_and_update_mux_table_entry(state_db, port_tbl, y_c
                             if y_cable_presence[0] is True and y_cable_asic_table is not None and mux_asic_table is not None and static_mux_asic_table is not None:
                                 # fill in the newly found entry
                                 read_y_cable_and_update_statedb_port_tbl(
-                                    logical_port_name, port_mapping, y_cable_tbl[asic_index])
+                                    logical_port_name, y_cable_tbl[asic_index])
                                 post_port_mux_info_to_db(
-                                    logical_port_name,  port_mapping, mux_tbl[asic_index])
+                                    logical_port_name,  mux_tbl[asic_index])
                                 post_port_mux_static_info_to_db(
-                                    logical_port_name,  port_mapping, static_tbl[asic_index])
+                                    logical_port_name,  static_tbl[asic_index])
 
                             else:
                                 # first create the state db y cable table and then fill in the entry
@@ -599,25 +612,25 @@ def check_identifier_presence_and_update_mux_table_entry(state_db, port_tbl, y_c
                                         state_db[asic_id], MUX_CABLE_INFO_TABLE)
                                 # fill the newly found entry
                                 read_y_cable_and_update_statedb_port_tbl(
-                                    logical_port_name, port_mapping, y_cable_tbl[asic_index])
+                                    logical_port_name, y_cable_tbl[asic_index])
                                 post_port_mux_info_to_db(
-                                    logical_port_name, port_mapping, mux_tbl[asic_index])
+                                    logical_port_name,  mux_tbl[asic_index])
                                 post_port_mux_static_info_to_db(
-                                    logical_port_name, port_mapping, static_tbl[asic_index])
+                                    logical_port_name,  static_tbl[asic_index])
                         else:
                             helper_logger.log_warning(
                                 "Error: Could not get transceiver info dict Y cable port {} while inserting entries".format(logical_port_name))
-                            create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, port_mapping)
+                            create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name)
 
                     else:
                         helper_logger.log_warning(
                             "Error: Could not establish transceiver presence for a Y cable port {} while inserting entries".format(logical_port_name))
-                        create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, port_mapping)
+                        create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name)
 
                 else:
                     helper_logger.log_warning(
                         "Error: Retreived multiple ports for a Y cable port {} while inserting entries".format(logical_port_name))
-                    create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, port_mapping)
+                    create_tables_and_insert_mux_unknown_entries(state_db, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name)
 
             else:
                 helper_logger.log_warning(
@@ -628,7 +641,7 @@ def check_identifier_presence_and_update_mux_table_entry(state_db, port_tbl, y_c
                 "Could not retreive state value inside mux_info_dict for {}, inside MUX_CABLE table".format(logical_port_name))
 
 
-def check_identifier_presence_and_delete_mux_table_entry(state_db, port_tbl, asic_index, logical_port_name, y_cable_presence, port_mapping, delete_change_event):
+def check_identifier_presence_and_delete_mux_table_entry(state_db, port_tbl, asic_index, logical_port_name, y_cable_presence, delete_change_event):
 
     y_cable_tbl = {}
     static_tbl, mux_tbl = {}, {}
@@ -662,7 +675,7 @@ def check_identifier_presence_and_delete_mux_table_entry(state_db, port_tbl, asi
                 delete_port_from_y_cable_table(logical_port_name, mux_tbl[asic_index])
                 delete_change_event[:] = [True]
                 # delete the y_cable instance
-                physical_port_list = port_mapping.logical_port_name_to_physical_port_list(logical_port_name)
+                physical_port_list = logical_port_name_to_physical_port_list(logical_port_name)
 
                 if len(physical_port_list) == 1:
 
@@ -674,7 +687,7 @@ def check_identifier_presence_and_delete_mux_table_entry(state_db, port_tbl, asi
                         "Error: Retreived multiple ports for a Y cable port {} while delete entries".format(logical_port_name))
 
 
-def init_ports_status_for_y_cable(platform_sfp, platform_chassis, y_cable_presence, port_mapping, stop_event=threading.Event()):
+def init_ports_status_for_y_cable(platform_sfp, platform_chassis, y_cable_presence, stop_event=threading.Event()):
     global y_cable_platform_sfputil
     global y_cable_platform_chassis
     global y_cable_port_instances
@@ -699,13 +712,13 @@ def init_ports_status_for_y_cable(platform_sfp, platform_chassis, y_cable_presen
         xcvrd_log_tbl[asic_id].set("Y_CABLE", fvs_updated)
 
     # Init PORT_STATUS table if ports are on Y cable
-    logical_port_list = port_mapping.logical_port_list
+    logical_port_list = y_cable_platform_sfputil.logical
     for logical_port_name in logical_port_list:
         if stop_event.is_set():
             break
 
         # Get the asic to which this port belongs
-        asic_index = port_mapping.get_asic_id_for_logical_port(
+        asic_index = y_cable_platform_sfputil.get_asic_id_for_logical_port(
             logical_port_name)
         if asic_index is None:
             helper_logger.log_warning(
@@ -714,7 +727,7 @@ def init_ports_status_for_y_cable(platform_sfp, platform_chassis, y_cable_presen
 
         if logical_port_name in port_table_keys[asic_index]:
             check_identifier_presence_and_update_mux_table_entry(
-                state_db, port_tbl, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, port_mapping, y_cable_presence)
+                state_db, port_tbl, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, y_cable_presence)
         else:
             # This port does not exist in Port table of config but is present inside
             # logical_ports after loading the port_mappings from port_config_file
@@ -723,7 +736,7 @@ def init_ports_status_for_y_cable(platform_sfp, platform_chassis, y_cable_presen
                 "Could not retreive port inside config_db PORT table {} for Y-Cable initiation".format(logical_port_name))
 
 
-def change_ports_status_for_y_cable_change_event(logical_port_dict, port_mapping, y_cable_presence, stop_event=threading.Event()):
+def change_ports_status_for_y_cable_change_event(port_dict, y_cable_presence, stop_event=threading.Event()):
     # Connect to CONFIG_DB and create port status table inside state_db
     config_db, state_db, port_tbl, y_cable_tbl = {}, {}, {}, {}
     static_tbl, mux_tbl = {}, {}
@@ -740,34 +753,41 @@ def change_ports_status_for_y_cable_change_event(logical_port_dict, port_mapping
         port_table_keys[asic_id] = port_tbl[asic_id].getKeys()
 
     # Init PORT_STATUS table if ports are on Y cable and an event is received
-    for logical_port_name, value in logical_port_dict.items():
-        # Get the asic to which this port belongs
-        asic_index = port_mapping.get_asic_id_for_logical_port(logical_port_name)
-        if asic_index is None:
-            helper_logger.log_warning(
-                "Got invalid asic index for {}, ignored".format(logical_port_name))
+    for key, value in port_dict.items():
+        if stop_event.is_set():
+            break
+        logical_port_list = y_cable_platform_sfputil.get_physical_to_logical(int(key))
+        if logical_port_list is None:
+            helper_logger.log_warning("Got unknown FP port index {}, ignored".format(key))
             continue
+        for logical_port_name in logical_port_list:
 
-        if logical_port_name in port_table_keys[asic_index]:
-            if value == sfp_status_helper.SFP_STATUS_INSERTED:
-                helper_logger.log_info("Got SFP inserted event")
-                check_identifier_presence_and_update_mux_table_entry(
-                    state_db, port_tbl, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, port_mapping, y_cable_presence)
-            elif value == sfp_status_helper.SFP_STATUS_REMOVED:
-                check_identifier_presence_and_delete_mux_table_entry(
-                    state_db, port_tbl, asic_index, logical_port_name, y_cable_presence, port_mapping, delete_change_event)
-            else:
-                try:
-                    # Now that the value is in bitmap format, let's convert it to number
-                    event_bits = int(value)
-                    if sfp_status_helper.is_error_block_eeprom_reading(event_bits):
-                        check_identifier_presence_and_delete_mux_table_entry(
-                            state_db, port_tbl, asic_index, logical_port_name, y_cable_presence, port_mapping, delete_change_event)
-                except:
-                    pass
-                # SFP return unkown event, just ignore for now.
-                helper_logger.log_warning("Got unknown event {}, ignored".format(value))
+            # Get the asic to which this port belongs
+            asic_index = y_cable_platform_sfputil.get_asic_id_for_logical_port(logical_port_name)
+            if asic_index is None:
+                helper_logger.log_warning("Got invalid asic index for {}, ignored".format(logical_port_name))
                 continue
+
+            if logical_port_name in port_table_keys[asic_index]:
+                if value == sfp_status_helper.SFP_STATUS_INSERTED:
+                    helper_logger.log_info("Got SFP inserted event")
+                    check_identifier_presence_and_update_mux_table_entry(
+                        state_db, port_tbl, y_cable_tbl, static_tbl, mux_tbl, asic_index, logical_port_name, y_cable_presence)
+                elif value == sfp_status_helper.SFP_STATUS_REMOVED:
+                    check_identifier_presence_and_delete_mux_table_entry(
+                        state_db, port_tbl, asic_index, logical_port_name, y_cable_presence, delete_change_event)
+                else:
+                    try:
+                        # Now that the value is in bitmap format, let's convert it to number
+                        event_bits = int(value)
+                        if sfp_status_helper.is_error_block_eeprom_reading(event_bits):
+                            check_identifier_presence_and_delete_mux_table_entry(
+                                state_db, port_tbl, asic_index, logical_port_name, y_cable_presence, delete_change_event)
+                    except:
+                        pass
+                    # SFP return unkown event, just ignore for now.
+                    helper_logger.log_warning("Got unknown event {}, ignored".format(value))
+                    continue
 
     # If there was a delete event and y_cable_presence was true, reaccess the y_cable presence
     if y_cable_presence[0] is True and delete_change_event[0] is True:
@@ -784,7 +804,8 @@ def change_ports_status_for_y_cable_change_event(logical_port_dict, port_mapping
                 break
 
 
-def delete_ports_status_for_y_cable(port_mapping):
+def delete_ports_status_for_y_cable():
+
     state_db, config_db, port_tbl, y_cable_tbl = {}, {}, {}, {}
     y_cable_tbl_keys = {}
     static_tbl, mux_tbl = {}, {}
@@ -803,11 +824,11 @@ def delete_ports_status_for_y_cable(port_mapping):
         port_tbl[asic_id] = swsscommon.Table(config_db[asic_id], "MUX_CABLE")
 
     # delete PORTS on Y cable table if ports on Y cable
-    logical_port_list = port_mapping.logical_port_list
+    logical_port_list = y_cable_platform_sfputil.logical
     for logical_port_name in logical_port_list:
 
         # Get the asic to which this port belongs
-        asic_index = port_mapping.get_asic_id_for_logical_port(
+        asic_index = y_cable_platform_sfputil.get_asic_id_for_logical_port(
             logical_port_name)
         if asic_index is None:
             helper_logger.log_warning(
@@ -818,7 +839,7 @@ def delete_ports_status_for_y_cable(port_mapping):
             delete_port_from_y_cable_table(logical_port_name, static_tbl[asic_index])
             delete_port_from_y_cable_table(logical_port_name, mux_tbl[asic_index])
             # delete the y_cable port instance
-            physical_port_list = port_mapping.logical_port_name_to_physical_port_list(logical_port_name)
+            physical_port_list = logical_port_name_to_physical_port_list(logical_port_name)
 
             if len(physical_port_list) == 1:
 
@@ -831,7 +852,7 @@ def delete_ports_status_for_y_cable(port_mapping):
                     "Error: Retreived multiple ports for a Y cable port {} while deleting entries".format(logical_port_name))
 
 
-def check_identifier_presence_and_update_mux_info_entry(state_db, mux_tbl, asic_index, logical_port_name, port_mapping):
+def check_identifier_presence_and_update_mux_info_entry(state_db, mux_tbl, asic_index, logical_port_name):
 
     # Get the namespaces in the platform
     config_db, port_tbl = {}, {}
@@ -852,11 +873,11 @@ def check_identifier_presence_and_update_mux_info_entry(state_db, mux_tbl, asic_
         mux_table_dict = dict(fvs)
         if "state" in mux_table_dict:
             val = mux_table_dict.get("state", None)
-            if val in ["active", "auto"]:
+            if val in ["active", "auto", "manual", "standby"]:
 
                 if mux_tbl.get(asic_index, None) is not None:
                     # fill in the newly found entry
-                    post_port_mux_info_to_db(logical_port_name, port_mapping, mux_tbl[asic_index])
+                    post_port_mux_info_to_db(logical_port_name,  mux_tbl[asic_index])
 
                 else:
                     # first create the state db y cable table and then fill in the entry
@@ -865,23 +886,53 @@ def check_identifier_presence_and_update_mux_info_entry(state_db, mux_tbl, asic_
                         asic_id = multi_asic.get_asic_index_from_namespace(namespace)
                         mux_tbl[asic_id] = swsscommon.Table(state_db[asic_id], MUX_CABLE_INFO_TABLE)
                     # fill the newly found entry
-                    post_port_mux_info_to_db(logical_port_name, port_mapping, mux_tbl[asic_index])
+                    post_port_mux_info_to_db(logical_port_name,  mux_tbl[asic_index])
             else:
                 helper_logger.log_warning(
                     "Could not retreive active or auto value for state kvp for {}, inside MUX_CABLE table".format(logical_port_name))
 
 
-def get_firmware_dict(physical_port, port_instance, target, side, mux_info_dict):
+def get_firmware_dict(physical_port, port_instance, target, side, mux_info_dict, logical_port_name):
 
     result = {}
-    if port_instance.download_firmware_status == port_instance.FIRMWARE_DOWNLOAD_STATUS_INPROGRESS or port_instance.download_firmware_status == port_instance.FIRMWARE_DOWNLOAD_STATUS_FAILED:
-        mux_info_dict[("version_{}_active".format(side))] = "N/A"
-        mux_info_dict[("version_{}_inactive".format(side))] = "N/A"
-        mux_info_dict[("version_{}_next".format(side))] = "N/A"
-        helper_logger.log_warning(
-            "trying to post firmware info while download in progress returning without execute {}".format(physical_port))
+    if port_instance.download_firmware_status == port_instance.FIRMWARE_DOWNLOAD_STATUS_INPROGRESS:
 
+        # if there is a firmware download in progress, retreive the last known firmware
+        state_db, mux_tbl = {}, {}
+        mux_firmware_dict = {}
+
+        namespaces = multi_asic.get_front_end_namespaces()
+        for namespace in namespaces:
+            asic_id = multi_asic.get_asic_index_from_namespace(namespace)
+            state_db[asic_id] = daemon_base.db_connect("STATE_DB", namespace)
+            mux_tbl[asic_id] = swsscommon.Table(
+                state_db[asic_id], MUX_CABLE_INFO_TABLE)
+
+        asic_index = y_cable_platform_sfputil.get_asic_id_for_logical_port(
+            logical_port_name)
+
+        (status, fvs) = mux_tbl[asic_index].get(logical_port_name)
+        if status is False:
+            helper_logger.log_warning("Could not retreive fieldvalue pairs for {}, inside state_db table {}".format(logical_port_name, mux_tbl[asic_index].getTableName()))
+            mux_info_dict[("version_{}_active".format(side))] = "N/A"
+            mux_info_dict[("version_{}_inactive".format(side))] = "N/A"
+            mux_info_dict[("version_{}_next".format(side))] = "N/A"
+            return
+
+        mux_firmware_dict = dict(fvs)
+
+        mux_info_dict[("version_{}_active".format(side))] = mux_firmware_dict.get(("version_{}_active".format(side)), None)
+        mux_info_dict[("version_{}_inactive".format(side))] = mux_firmware_dict.get(("version_{}_inactive".format(side)), None)
+        mux_info_dict[("version_{}_next".format(side))] = mux_firmware_dict.get(("version_{}_next".format(side)), None)
+
+        helper_logger.log_warning(
+            "trying to get/post firmware info while download in progress returning with last known firmware without execute {}".format(physical_port))
         return
+
+    elif port_instance.download_firmware_status == port_instance.FIRMWARE_DOWNLOAD_STATUS_FAILED:
+        # if there is a firmware download failed, retreive the current MCU's firmware with a log message
+        helper_logger.log_error(
+            "Firmware Download API failed in the previous run, firmware download status was set to failed;retry required {}".format(physical_port))
 
     with y_cable_port_locks[physical_port]:
         try:
@@ -901,7 +952,7 @@ def get_firmware_dict(physical_port, port_instance, target, side, mux_info_dict)
         mux_info_dict[("version_{}_next".format(side))] = "N/A"
 
 
-def get_muxcable_info(physical_port, logical_port_name, port_mapping):
+def get_muxcable_info(physical_port, logical_port_name):
 
     mux_info_dict = {}
     y_cable_tbl, state_db = {}, {}
@@ -917,7 +968,7 @@ def get_muxcable_info(physical_port, logical_port_name, port_mapping):
         state_db[asic_id] = daemon_base.db_connect("STATE_DB", namespace)
         y_cable_tbl[asic_id] = swsscommon.Table(state_db[asic_id], swsscommon.STATE_HW_MUX_CABLE_TABLE_NAME)
 
-    asic_index = port_mapping.get_asic_id_for_logical_port(
+    asic_index = y_cable_platform_sfputil.get_asic_id_for_logical_port(
         logical_port_name)
     if asic_index is None:
         helper_logger.log_warning("Got invalid asic index for {}, ignored".format(logical_port_name))
@@ -1087,13 +1138,13 @@ def get_muxcable_info(physical_port, logical_port_name, port_mapping):
         else:
             mux_info_dict["link_status_nic"] = "down"
 
-    get_firmware_dict(physical_port, port_instance, port_instance.TARGET_NIC, "nic", mux_info_dict)
+    get_firmware_dict(physical_port, port_instance, port_instance.TARGET_NIC, "nic", mux_info_dict, logical_port_name)
     if read_side == 1:
-        get_firmware_dict(physical_port, port_instance, port_instance.TARGET_TOR_A, "self", mux_info_dict)
-        get_firmware_dict(physical_port, port_instance, port_instance.TARGET_TOR_B, "peer", mux_info_dict)
+        get_firmware_dict(physical_port, port_instance, port_instance.TARGET_TOR_A, "self", mux_info_dict, logical_port_name)
+        get_firmware_dict(physical_port, port_instance, port_instance.TARGET_TOR_B, "peer", mux_info_dict, logical_port_name)
     else:
-        get_firmware_dict(physical_port, port_instance, port_instance.TARGET_TOR_A, "peer", mux_info_dict)
-        get_firmware_dict(physical_port, port_instance, port_instance.TARGET_TOR_B, "self", mux_info_dict)
+        get_firmware_dict(physical_port, port_instance, port_instance.TARGET_TOR_A, "peer", mux_info_dict, logical_port_name)
+        get_firmware_dict(physical_port, port_instance, port_instance.TARGET_TOR_B, "self", mux_info_dict, logical_port_name)
 
     with y_cable_port_locks[physical_port]:
         try:
@@ -1146,7 +1197,7 @@ def get_muxcable_info(physical_port, logical_port_name, port_mapping):
     return mux_info_dict
 
 
-def get_muxcable_static_info(physical_port, logical_port_name, port_mapping):
+def get_muxcable_static_info(physical_port, logical_port_name):
 
     mux_static_info_dict = {}
     y_cable_tbl, state_db = {}, {}
@@ -1163,7 +1214,7 @@ def get_muxcable_static_info(physical_port, logical_port_name, port_mapping):
         y_cable_tbl[asic_id] = swsscommon.Table(
             state_db[asic_id], swsscommon.STATE_HW_MUX_CABLE_TABLE_NAME)
 
-    asic_index = port_mapping.get_asic_id_for_logical_port(
+    asic_index = y_cable_platform_sfputil.get_asic_id_for_logical_port(
         logical_port_name)
     if asic_index is None:
         helper_logger.log_warning(
@@ -1260,9 +1311,9 @@ def get_muxcable_static_info(physical_port, logical_port_name, port_mapping):
     return mux_static_info_dict
 
 
-def post_port_mux_info_to_db(logical_port_name, port_mapping, table):
+def post_port_mux_info_to_db(logical_port_name, table):
 
-    physical_port_list = port_mapping.logical_port_name_to_physical_port_list(logical_port_name)
+    physical_port_list = logical_port_name_to_physical_port_list(logical_port_name)
     if physical_port_list is None:
         helper_logger.log_error("No physical ports found for logical port '{}'".format(logical_port_name))
         return -1
@@ -1277,7 +1328,7 @@ def post_port_mux_info_to_db(logical_port_name, port_mapping, table):
             helper_logger.log_warning("Error: trying to post mux info without presence of port {}".format(logical_port_name))
             continue
 
-        mux_info_dict = get_muxcable_info(physical_port, logical_port_name, port_mapping)
+        mux_info_dict = get_muxcable_info(physical_port, logical_port_name)
         if mux_info_dict is not None and mux_info_dict is not -1:
             #transceiver_dict[physical_port] = port_info_dict
             fvs = swsscommon.FieldValuePairs(
@@ -1313,9 +1364,10 @@ def post_port_mux_info_to_db(logical_port_name, port_mapping, table):
             return -1
 
 
-def post_port_mux_static_info_to_db(logical_port_name, port_mapping, static_table):
+def post_port_mux_static_info_to_db(logical_port_name, static_table):
 
-    physical_port_list = port_mapping.logical_port_name_to_physical_port_list(logical_port_name)
+    physical_port_list = logical_port_name_to_physical_port_list(
+        logical_port_name)
     if physical_port_list is None:
         helper_logger.log_error("No physical ports found for logical port '{}'".format(logical_port_name))
         return -1
@@ -1330,7 +1382,7 @@ def post_port_mux_static_info_to_db(logical_port_name, port_mapping, static_tabl
         if not y_cable_wrapper_get_presence(physical_port):
             continue
 
-        mux_static_info_dict = get_muxcable_static_info(physical_port, logical_port_name, port_mapping)
+        mux_static_info_dict = get_muxcable_static_info(physical_port, logical_port_name)
 
         if mux_static_info_dict is not None and mux_static_info_dict is not -1:
             #transceiver_dict[physical_port] = port_info_dict
@@ -1372,7 +1424,7 @@ def post_port_mux_static_info_to_db(logical_port_name, port_mapping, static_tabl
             return -1
 
 
-def post_mux_static_info_to_db(is_warm_start, port_mapping, stop_event=threading.Event()):
+def post_mux_static_info_to_db(is_warm_start, stop_event=threading.Event()):
     # Connect to STATE_DB and create transceiver mux/static info tables
     state_db, static_tbl = {}, {}
 
@@ -1385,20 +1437,21 @@ def post_mux_static_info_to_db(is_warm_start, port_mapping, stop_event=threading
             state_db[asic_id], MUX_CABLE_STATIC_INFO_TABLE)
 
     # Post all the current interface dom/sfp info to STATE_DB
-    logical_port_list = port_mapping.logical_port_list
+    logical_port_list = y_cable_platform_sfputil.logical
     for logical_port_name in logical_port_list:
         if stop_event.is_set():
             break
 
         # Get the asic to which this port belongs
-        asic_index = port_mapping.get_asic_id_for_logical_port(logical_port_name)
+        asic_index = y_cable_platform_sfputil.get_asic_id_for_logical_port(
+            logical_port_name)
         if asic_index is None:
             helper_logger.log_warning("Got invalid asic index for {}, ignored".format(logical_port_name))
             continue
-        post_port_mux_static_info_to_db(logical_port_name, port_mapping, mux_tbl[asic_index])
+        post_port_mux_static_info_to_db(logical_port_name, mux_tbl[asic_index])
 
 
-def post_mux_info_to_db(is_warm_start, port_mapping, stop_event=threading.Event()):
+def post_mux_info_to_db(is_warm_start, stop_event=threading.Event()):
     # Connect to STATE_DB and create transceiver mux/static info tables
     state_db, mux_tbl, static_tbl = {}, {}, {}
 
@@ -1411,24 +1464,26 @@ def post_mux_info_to_db(is_warm_start, port_mapping, stop_event=threading.Event(
             state_db[asic_id], MUX_CABLE_INFO_TABLE)
 
     # Post all the current interface dom/sfp info to STATE_DB
-    logical_port_list = port_mapping.logical_port_list
+    logical_port_list = y_cable_platform_sfputil.logical
     for logical_port_name in logical_port_list:
         if stop_event.is_set():
             break
 
         # Get the asic to which this port belongs
-        asic_index = port_mapping.get_asic_id_for_logical_port(logical_port_name)
+        asic_index = y_cable_platform_sfputil.get_asic_id_for_logical_port(
+            logical_port_name)
         if asic_index is None:
             helper_logger.log_warning(
                 "Got invalid asic index for {}, ignored".format(logical_port_name))
             continue
-        post_port_mux_info_to_db(logical_port_name,  port_mapping, mux_tbl[asic_index])
+        post_port_mux_info_to_db(logical_port_name,  mux_tbl[asic_index])
 
 
 def task_download_firmware_worker(port, physical_port, port_instance, file_full_path, xcvrd_down_fw_rsp_tbl, xcvrd_down_fw_cmd_sts_tbl, rc):
     helper_logger.log_debug("worker thread launched for downloading physical port {} path {}".format(physical_port, file_full_path))
     try:
         status = port_instance.download_firmware(file_full_path)
+        time.sleep(5)
     except Exception as e:
         status = -1
         helper_logger.log_warning("Failed to execute the download firmware API for port {} due to {}".format(physical_port,repr(e)))
@@ -1440,12 +1495,11 @@ def task_download_firmware_worker(port, physical_port, port_instance, file_full_
 
 # Thread wrapper class to update y_cable status periodically
 class YCableTableUpdateTask(object):
-    def __init__(self, port_mapping):
+    def __init__(self):
         self.task_thread = None
         self.task_cli_thread = None
         self.task_download_firmware_thread = {}
         self.task_stopping_event = threading.Event()
-        self.port_mapping = copy.deepcopy(port_mapping)
 
         if multi_asic.is_multi_asic():
             # Load the namespace details first from the database_global.json file.
@@ -1458,7 +1512,6 @@ class YCableTableUpdateTask(object):
         y_cable_tbl_keys = {}
         mux_cable_command_tbl, y_cable_command_tbl = {}, {}
         mux_metrics_tbl = {}
-        asic_context = {}
 
         sel = swsscommon.Select()
 
@@ -1481,11 +1534,8 @@ class YCableTableUpdateTask(object):
             mux_metrics_tbl[asic_id] = swsscommon.Table(
                 state_db[asic_id], swsscommon.STATE_MUX_METRICS_TABLE_NAME)
             y_cable_tbl_keys[asic_id] = y_cable_tbl[asic_id].getKeys()
-            port_tbl = swsscommon.SubscriberStateTable(config_db[asic_id], swsscommon.CFG_PORT_TABLE_NAME)
-            asic_context[port_tbl] = asic_id
             sel.addSelectable(status_tbl[asic_id])
             sel.addSelectable(mux_cable_command_tbl[asic_id])
-            sel.addSelectable(port_tbl)
 
         # Listen indefinitely for changes to the HW_MUX_CABLE_TABLE in the Application DB's
         while True:
@@ -1511,7 +1561,6 @@ class YCableTableUpdateTask(object):
             # Get the corresponding namespace from redisselect db connector object
             namespace = redisSelectObj.getDbConnector().getNamespace()
             asic_index = multi_asic.get_asic_index_from_namespace(namespace)
-            read_port_config_change(asic_context, self.port_mapping, helper_logger, self.port_mapping.handle_port_change_event)
 
             while True:
                 (port, op, fvp) = status_tbl[asic_index].pop()
@@ -1545,7 +1594,7 @@ class YCableTableUpdateTask(object):
                         read_side = mux_port_dict.get("read_side")
                         # Now whatever is the state requested, toggle the mux appropriately
                         helper_logger.log_debug("Y_CABLE_DEBUG: xcvrd trying to transition port {} from {} to {}".format(port, old_status, new_status))
-                        active_side = update_tor_active_side(read_side, new_status, port, self.port_mapping)
+                        active_side = update_tor_active_side(read_side, new_status, port)
                         if active_side == -1:
                             helper_logger.log_warning("ERR: Got a change event for toggle but could not toggle the mux-direction for port {} state from {} to {}, writing unknown".format(
                                 port, old_status, new_status))
@@ -1593,7 +1642,7 @@ class YCableTableUpdateTask(object):
                                 continue
                             mux_port_dict = dict(fv)
                             read_side = mux_port_dict.get("read_side")
-                            update_appdb_port_mux_cable_response_table(port_m, self.port_mapping, asic_index, appl_db, int(read_side))
+                            update_appdb_port_mux_cable_response_table(port_m, asic_index, appl_db, int(read_side))
 
 
     def task_cli_worker(self):
@@ -1751,7 +1800,7 @@ class YCableTableUpdateTask(object):
 
                     if "state" in fvp_dict:
 
-                        physical_port = get_ycable_physical_port_from_logical_port(port, self.port_mapping)
+                        physical_port = get_ycable_physical_port_from_logical_port(port)
                         if physical_port is None or physical_port == PHYSICAL_PORT_MAPPING_ERROR:
                             state = 'cable not present'
                             # error scenario update table accordingly
@@ -1760,7 +1809,7 @@ class YCableTableUpdateTask(object):
                             set_result_and_delete_port('state', state, xcvrd_show_hwmode_dir_cmd_sts_tbl[asic_index], xcvrd_show_hwmode_dir_rsp_tbl[asic_index], port)
                             break
 
-                        port_instance = get_ycable_port_instance_from_logical_port(port, self.port_mapping)
+                        port_instance = get_ycable_port_instance_from_logical_port(port)
                         if port_instance is None or port_instance in port_mapping_error_values:
                             # error scenario update table accordingly
                             state = 'not Y-Cable port'
@@ -1829,7 +1878,7 @@ class YCableTableUpdateTask(object):
                         config_state = str(fvp_dict["config"])
 
                         status = 'False'
-                        physical_port = get_ycable_physical_port_from_logical_port(port, self.port_mapping)
+                        physical_port = get_ycable_physical_port_from_logical_port(port)
                         if physical_port is None or physical_port == PHYSICAL_PORT_MAPPING_ERROR:
                             # error scenario update table accordingly
                             helper_logger.log_error(
@@ -1837,7 +1886,7 @@ class YCableTableUpdateTask(object):
                             set_result_and_delete_port('result', status, xcvrd_config_hwmode_state_cmd_sts_tbl[asic_index], xcvrd_config_hwmode_state_rsp_tbl[asic_index], port)
                             break
 
-                        port_instance = get_ycable_port_instance_from_logical_port(port, self.port_mapping)
+                        port_instance = get_ycable_port_instance_from_logical_port(port)
                         if port_instance is None or port_instance in port_mapping_error_values:
                             # error scenario update table accordingly
                             helper_logger.log_error(
@@ -1915,7 +1964,7 @@ class YCableTableUpdateTask(object):
                     if "state" in fvp_dict:
 
                         state = 'unknown'
-                        physical_port = get_ycable_physical_port_from_logical_port(port, self.port_mapping)
+                        physical_port = get_ycable_physical_port_from_logical_port(port)
                         if physical_port is None or physical_port == PHYSICAL_PORT_MAPPING_ERROR:
                             # error scenario update table accordingly
                             helper_logger.log_error(
@@ -1976,7 +2025,7 @@ class YCableTableUpdateTask(object):
                         config_mode = str(fvp_dict["config"])
 
                         status = 'False'
-                        physical_port = get_ycable_physical_port_from_logical_port(port, self.port_mapping)
+                        physical_port = get_ycable_physical_port_from_logical_port(port)
                         if physical_port is None or physical_port == PHYSICAL_PORT_MAPPING_ERROR:
                             # error scenario update table accordingly
                             helper_logger.log_error(
@@ -1984,7 +2033,7 @@ class YCableTableUpdateTask(object):
                             set_result_and_delete_port('result', status, xcvrd_config_hwmode_swmode_cmd_sts_tbl[asic_index], xcvrd_config_hwmode_swmode_rsp_tbl[asic_index], port)
                             break
 
-                        port_instance = get_ycable_port_instance_from_logical_port(port, self.port_mapping)
+                        port_instance = get_ycable_port_instance_from_logical_port(port)
                         if port_instance is None or port_instance in port_mapping_error_values:
                             # error scenario update table accordingly
                             helper_logger.log_error(
@@ -2062,7 +2111,7 @@ class YCableTableUpdateTask(object):
                             set_result_and_delete_port('status', status, xcvrd_down_fw_cmd_sts_tbl[asic_index], xcvrd_down_fw_rsp_tbl[asic_index], port)
                             break
 
-                        physical_port = get_ycable_physical_port_from_logical_port(port, self.port_mapping)
+                        physical_port = get_ycable_physical_port_from_logical_port(port)
                         if physical_port is None or physical_port == PHYSICAL_PORT_MAPPING_ERROR:
                             # error scenario update table accordingly
                             helper_logger.log_error(
@@ -2070,7 +2119,7 @@ class YCableTableUpdateTask(object):
                             set_result_and_delete_port('status', status, xcvrd_down_fw_cmd_sts_tbl[asic_index], xcvrd_down_fw_rsp_tbl[asic_index], port)
                             break
 
-                        port_instance = get_ycable_port_instance_from_logical_port(port, self.port_mapping)
+                        port_instance = get_ycable_port_instance_from_logical_port(port)
                         if port_instance is None or port_instance in port_mapping_error_values:
                             # error scenario update table accordingly
                             helper_logger.log_error(
@@ -2111,7 +2160,7 @@ class YCableTableUpdateTask(object):
 
 
                         status = 'False'
-                        physical_port = get_ycable_physical_port_from_logical_port(port, self.port_mapping)
+                        physical_port = get_ycable_physical_port_from_logical_port(port)
                         if physical_port is None or physical_port == PHYSICAL_PORT_MAPPING_ERROR:
                             # error scenario update table accordingly
                             helper_logger.log_warning("Error: Could not get physical port for cli cmd show firmware port {}".format(port))
@@ -2119,7 +2168,7 @@ class YCableTableUpdateTask(object):
                             set_show_firmware_fields(port, mux_info_dict, xcvrd_show_fw_res_tbl[asic_index])
                             break
 
-                        port_instance = get_ycable_port_instance_from_logical_port(port, self.port_mapping)
+                        port_instance = get_ycable_port_instance_from_logical_port(port)
                         if port_instance is None or port_instance in port_mapping_error_values:
                             # error scenario update table accordingly
                             helper_logger.log_warning("Error: Could not get port instance for cli cmd show firmware command port {}".format(port))
@@ -2142,13 +2191,13 @@ class YCableTableUpdateTask(object):
                             break
 
 
-                        get_firmware_dict(physical_port, port_instance, port_instance.TARGET_NIC, "nic", mux_info_dict)
+                        get_firmware_dict(physical_port, port_instance, port_instance.TARGET_NIC, "nic", mux_info_dict, port)
                         if read_side == port_instance.TARGET_TOR_A:
-                            get_firmware_dict(physical_port, port_instance, port_instance.TARGET_TOR_A, "self", mux_info_dict)
-                            get_firmware_dict(physical_port, port_instance, port_instance.TARGET_TOR_B, "peer", mux_info_dict)
+                            get_firmware_dict(physical_port, port_instance, port_instance.TARGET_TOR_A, "self", mux_info_dict, port)
+                            get_firmware_dict(physical_port, port_instance, port_instance.TARGET_TOR_B, "peer", mux_info_dict, port)
                         else:
-                            get_firmware_dict(physical_port, port_instance, port_instance.TARGET_TOR_A, "peer", mux_info_dict)
-                            get_firmware_dict(physical_port, port_instance, port_instance.TARGET_TOR_B, "self", mux_info_dict)
+                            get_firmware_dict(physical_port, port_instance, port_instance.TARGET_TOR_A, "peer", mux_info_dict, port)
+                            get_firmware_dict(physical_port, port_instance, port_instance.TARGET_TOR_B, "self", mux_info_dict, port)
 
                         status = 'True'
                         set_show_firmware_fields(port, mux_info_dict, xcvrd_show_fw_res_tbl[asic_index])
@@ -2184,14 +2233,14 @@ class YCableTableUpdateTask(object):
                                 break
 
 
-                        physical_port = get_ycable_physical_port_from_logical_port(port, self.port_mapping)
+                        physical_port = get_ycable_physical_port_from_logical_port(port)
                         if physical_port is None or physical_port == PHYSICAL_PORT_MAPPING_ERROR:
                             # error scenario update table accordingly
                             helper_logger.log_warning("Error: Could not get physical port for cli cmd mux activate firmware port {}".format(port))
                             set_result_and_delete_port('status', status, xcvrd_acti_fw_cmd_sts_tbl[asic_index], xcvrd_acti_fw_rsp_tbl[asic_index], port)
                             break
 
-                        port_instance = get_ycable_port_instance_from_logical_port(port, self.port_mapping)
+                        port_instance = get_ycable_port_instance_from_logical_port(port)
                         if port_instance is None or port_instance in port_mapping_error_values:
                             helper_logger.log_warning("Error: Could not get port instance for cli cmd mux activate firmware port {}".format(port))
                             # error scenario update table accordingly
@@ -2237,14 +2286,14 @@ class YCableTableUpdateTask(object):
 
 
 
-                        physical_port = get_ycable_physical_port_from_logical_port(port, self.port_mapping)
+                        physical_port = get_ycable_physical_port_from_logical_port(port)
                         if physical_port is None or physical_port == PHYSICAL_PORT_MAPPING_ERROR:
                             # error scenario update table accordingly
                             helper_logger.log_warning("Error: Could not get physical port for cli cmd mux rollback firmware port {}".format(port))
                             set_result_and_delete_port('status', status, xcvrd_roll_fw_cmd_sts_tbl[asic_index], xcvrd_roll_fw_rsp_tbl[asic_index], port)
                             break
 
-                        port_instance = get_ycable_port_instance_from_logical_port(port, self.port_mapping)
+                        port_instance = get_ycable_port_instance_from_logical_port(port)
                         if port_instance is None or port_instance in port_mapping_error_values:
                             # error scenario update table accordingly
                             helper_logger.log_warning("Error: Could not get port instance for cli cmd mux rollback firmware port {}".format(port))
